moved to map screen
made map screen
